move assign_hash_ids from test_utils to unit_utils
